# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Access API",
        "description": "API endpoints for UGV-Based Disaster Management System",
        "version": "1.0"
    },
    "basePath": "/",
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Bearer token"
        }
    },
    "paths": {
...
        "/api/devices": {
            "post": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Registration",
                "description": "Register a new device\nName and MAC Address must be unique",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Device data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Device1",
                                    "required": true
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16",
                                    "required": true
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Device registered successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device registered successfully."
                                },
                                "device_id": {
                                    "type": "string",
                                    "example": "6055a0ae20e90e08641e22ef"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "409": {
                        "description": "Conflict - A device with the same MAC address is already registered"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/{device_id}": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Info Retrievement [ADMIN]",
                "description": "Get Device Information for Admin interface",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to retrieve",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON object of device information",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "device_id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Robot1"
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                },
                                "status": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid token"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Info Update [ADMIN]",
                "description": "update name, mac, or type for the Device\n\nDevice Types:\n1: UGV\n2: UAV\n3: DOG\n4: CHARGING_STATION\n 5: BROKER",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to update",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Device data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "new_name"
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 2
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device information is updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device information is updated successfully."
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Robot1"
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Deactivation [ADMIN]",
                "description": "Deactivate a device",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to deactivate",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device is deactivated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device is deactivated successfully."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/all": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device List Retrievement [ADMIN]",
                "description": "Get list of devices with pagenation for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page-number",
                        "schema": {
                            "type": "integer"
                        },
                        "required": false,
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "in": "query",
                        "name": "page-size",
                        "schema": {
                            "type": "integer"
                        },
                        "required": false,
                        "description": "Number of items per page",
                        "default": 5
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "integer"
                        },
                        "required": false,
                        "description": "filter for the device status"
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "type": "integer"
                        },
                        "required": false,
                        "description": "filter for the device type"
                    },
                    {
                        "in": "query",
                        "name": "mission",
                        "schema": {
                            "type": "integer"
                        },
                        "required": false,
                        "description": "[mission_id] devices for mission to include in the list"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "6055a0ae80e90e08641e22ef"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Robot1"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    }
}